% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nucleoSimShiny.R
\docType{package}
\name{nucleoSimShiny-package}
\alias{nucleoSimShiny}
\alias{nucleoSimShiny-package}
\title{nucleoSimShiny: Interactive Generator of Nucleosomes Maps}
\description{
This package provides a graphical user interface (GUI)  to the nucleoSim 
package which generates nucleosomes maps.The GUI implementation relies on
an shiny framework.
}
\author{
Astrid Deschenes

Maintainer:
Astrid Deschenes <astrid-louise.deschenes@crchudequebec.ulaval.ca>
}
\seealso{
\itemize{
\item \code{\link{startNucleoSimShiny}} { for launching an interactive 
session}
}
}
\keyword{package}

