% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyNucleoSim.R
\docType{package}
\name{shinyNucleoSim-package}
\alias{shinyNucleoSim}
\alias{shinyNucleoSim-package}
\title{shinyNucleoSim: Interactive Generator of Nucleosomes Maps}
\description{
This package provides a graphical user interface (GUI)  to the nucleoSim 
package which generates nucleosomes maps.The GUI implementation relies on
an shiny framework.
}
\author{
Astrid Deschenes

Maintainer:
Astrid Deschenes <astrid-louise.deschenes@crchudequebec.ulaval.ca>
}
\seealso{
\itemize{
\item \code{\link{startShinyNucleoSim}} { for launching an interactive 
session}
}
}
\keyword{package}

