% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nucleoSimShiny-internal.R
\name{startNucleoSimShiny}
\alias{startNucleoSimShiny}
\title{Launch interactive session}
\usage{
startNucleoSimShiny()
}
\description{
Start an interactive session which enable users to generate 
nucleosomes maps through an interface. The implementation relies on
the \code{shiny} framework.
}
\examples{
if (interactive()) {
    startNucleoSimShiny()
}

}
\seealso{
\itemize{ 
\item \code{\link[shiny]{runApp}}
}
}

