% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyNucleoSim-internal.R
\name{startShinyNucleoSim}
\alias{startShinyNucleoSim}
\title{Launch interactive session}
\usage{
startShinyNucleoSim()
}
\description{
Start shiny session enabling users to generate nucleosomes
maps through an interface.
}
\examples{
if (interactive()) {
    startShinyNucleoSim()
}

}
\seealso{
\itemize{ 
\item \code{\link[shiny]{runApp}}
}
}

