\name{shinyNucleoSim-package}
\alias{shinyNucleoSim-package}
\alias{shinyNucleoSim}
\docType{package}
\title{
\packageTitle{shinyNucleoSim}
}
\description{
\packageDescription{shinyNucleoSim}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{shinyNucleoSim}
\packageIndices{shinyNucleoSim}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{shinyNucleoSim}

Maintainer: \packageMaintainer{shinyNucleoSim}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
