% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyNucleoSim.R
\docType{package}
\name{shinyNucleoSim-package}
\alias{shinyNucleoSim}
\alias{shinyNucleoSim-package}
\title{shinyNucleoSim: Interactive Generator of Nucleosomes Maps}
\description{
This package TODO
}
\author{
Astrid Deschenes

Maintainer:
Astrid Deschenes <astrid-louise.deschenes@crchudequebec.ulaval.ca>
}
\seealso{
\itemize{
\item \code{\link{startShinyNucleoSim}} { TODO}
}
}
\keyword{package}

